refactor(agents): extract agents channel and track-session script template
Move Output channel ownership out of resume-hint and the POSIX script body
into dedicated modules so the catalog stays data-focused.
